Add files via upload
This visualization consists of a bar graph showing every Washington county on the x-axis and the number of hot days on the y-axis. This visualization can show a trend on which counties have the most hot days and we can paint an accurate picture on what counties have the most and which have the least. Also used the 'alpha' function for the formation to show the lowest amount for each census tract and the highest by having overlapping bar graphs for each county.